conf cleaning, pom cleaning
