+++ airport-rest/pom.xml
@@ -26,6 +26,11 @@
             <artifactId>airport-dao</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
+<!--        <dependency>
+            <groupId>cz.muni</groupId>
+            <artifactId>airport-mvc</artifactId>
+            <version>${project.parent.version}</version>
+        </dependency>-->
 
 
         <dependency>
@@ -75,7 +80,7 @@
             <groupId>org.springframework.hateoas</groupId>
             <artifactId>spring-hateoas</artifactId>
         </dependency>
-        <dependency>
+<!--        <dependency>
             <groupId>com.jayway.jsonpath</groupId>
             <artifactId>json-path</artifactId>
         </dependency>
@@ -92,10 +97,10 @@
             <groupId>com.github.dandelion</groupId>
             <artifactId>datatables-thymeleaf</artifactId>
             <version>1.1.0</version>
-        </dependency>
+        </dependency>-->
 
         <!-- Thymeleaf -->
-        <dependency>
+<!--        <dependency>
             <groupId>org.thymeleaf</groupId>
             <artifactId>thymeleaf</artifactId>
             <scope>compile</scope>
@@ -112,7 +117,7 @@
         <dependency>
             <groupId>org.thymeleaf.extras</groupId>
             <artifactId>thymeleaf-extras-springsecurity4</artifactId>
-        </dependency>
+        </dependency>-->
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-thymeleaf</artifactId>
@@ -129,7 +134,7 @@
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-log4j</artifactId>
         </dependency>
-        <dependency>
+<!--        <dependency>
             <groupId>commons-fileupload</groupId>
             <artifactId>commons-fileupload</artifactId>
         </dependency>
@@ -137,7 +142,7 @@
             <groupId>org.thymeleaf.extras</groupId>
             <artifactId>thymeleaf-extras-java8time</artifactId>
             <version>2.1.0.RELEASE</version>
-        </dependency>
+        </dependency>-->
         <dependency>
             <groupId>joda-time</groupId>
             <artifactId>joda-time</artifactId>
